Refactor GalleryListPage and related components
- Move GalleryListPage from pages to components folder
- Update import paths in DisplayCategoryListWrap and ProductSectionListWrap
