+++ src/components/GalleryList/ProductItem.jsx
@@ -3,8 +3,8 @@ import { Link } from 'react-router-dom';
 import { ThumbItem } from '@shopby/react-components';
 import { calculateDiscountedPrice } from '@shopby/shared';
 
-import ProductThumbBadge from '../../components/ProductThumbBadge';
-import ProductThumbInfo from '../../components/ProductThumbInfo';
+import ProductThumbBadge from '../ProductThumbBadge';
+import ProductThumbInfo from '../ProductThumbInfo';
 
 const ProductItem = ({
   productNo,
