+++ src/components/GalleryList/index.js
@@ -1,0 +1,7 @@
+import GalleryList from './GalleryList';
+import ProductList from './ProductList';
+import ProductItem from './ProductItem';
+import NoSearchProduct from './NoSearchProduct';
+
+export { GalleryList, ProductList, ProductItem, NoSearchProduct };
+export default GalleryList;
