--- src/pages/GalleryListPage/index.js
@@ -1,7 +1,0 @@
-import GalleryListPage from './GalleryListPage';
-import ProductList from './ProductList';
-import ProductItem from './ProductItem';
-import NoSearchProduct from './NoSearchProduct';
-
-export { GalleryListPage, ProductList, ProductItem, NoSearchProduct };
-export default GalleryListPage;
